extract a CMake module responsible for initializing global variables, e.g. Config and Context
some free methods are using leaking variants  for rust test suite
