+++ src/module-init/config.c
